refactor(websocket): update to use the new structure
